requirements.txt
# Core
python-telegram-bot==20.8
python-dotenv==1.0.0

# Database
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
redis==5.0.1

# Media Processing
yt-dlp==2023.12.30
ffmpeg-python==0.2.0
python-multipart==0.0.6

# Utilities
pydantic==2.5.3
python-jose==3.3.0
passlib==1.7.4
python-dateutil==2.8.2

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
mypy==1.7.0

# Documentation
mkdocs==1.5.3
mkdocs-material==9.4.1
